Split functions into lib
